(original rev 927) View2d just needs some clean-up this will be the first in a series of commits . . . just cleaning stuff up . . . explicitly stating casts when required . . . etc.
git-svn-id: file:///tmp/current/omni/branches/staging@708 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/view2d/omView2d.h
@@ -228,7 +228,7 @@ private:
 	OmId mSecondSegId;
 	
 	int mTileCount;
-	Vector2f mMousePoint;
+	Vector2i mMousePoint;
 	
 	// OmCamera2d mCamera;
 	OmGenericManager< OmView2dWidget > mView2dWidgetManager;
